variance of laplacian format changed

+++ lib/VarianceOfLaplacian/VarianceOfLaplacian.cpp
@@ -49,9 +49,9 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[])
     // Convert mxArray inputs into OpenCV types
     cv::Ptr<cv::Mat> imgCV = ocvMxArrayToImage_uint8(prhs[0], true);
     
-    cv::Mat imgLap(imgCV->rows, imgCV->cols, CV_8U);
+    cv::Mat imgLap(imgCV->rows, imgCV->cols, CV_32F);
     
-    cv::Laplacian(*imgCV, imgLap, CV_8U, 3);
+    cv::Laplacian(*imgCV, imgLap, CV_32F, 3);
     cv::Mat mean, stddev;
     cv::meanStdDev(imgLap, mean, stddev);
     cv::Mat variance(1,1,CV_32F);
